Fixed compilation warning

+++ src/RouteManager.cpp
@@ -362,7 +362,7 @@ void RouteManager::calculateCurrentSplitTransformation(double currentTime, doubl
 		// check if we are inside the time range of two consecutive controls
 		if (runnerOffsetTime >= firstSplitOffsetTime && runnerOffsetTime < secondSplitOffsetTime)
 		{
-			if (i >= defaultRoute.splitTransformations.size())
+			if (i >= (int)defaultRoute.splitTransformations.size())
 				break;
 
 			if (instantTransitionRequested)
